ml-pipeline/requirements.txt
torch==2.1.2
torchvision==0.16.2
ultralytics==8.1.0
monai==1.3.0
firebase-admin==6.4.0
opencv-python==4.9.0.80
numpy==1.26.3
pillow==10.2.0
SimpleITK==2.3.1
scikit-image==0.22.0
nibabel==5.2.0
scipy==1.11.4

